settings decomposition
